LayoutTests/platform/mac-wk2/imported/w3c/web-platform-tests/mathml/relations/css-styling/displaystyle-3-expected.txt


PASS movablelimits for munder element (displaystyle=false) 
FAIL movablelimits for munder element (displaystyle=true) assert_approx_equals: expected 296.375 +/- 1 but got 320.375
PASS movablelimits for mover element (displaystyle=false) 
FAIL movablelimits for mover element (displaystyle=true) assert_approx_equals: expected 375.46875 +/- 1 but got 399.46875
PASS movablelimits for munderover element (displaystyle=false) 
FAIL movablelimits for munderover element (displaystyle=true) assert_approx_equals: expected 454.5625 +/- 1 but got 478.5625
XX
X
 
XX
X
 
XX
X
X
 
XX
X
 
XX
X
 
XX
X
X
